refactor: move AudioPlayerService into its own service folder, compare device format directly
